refactor(tools): move topological_sort from tools.utils to db.sort
KAG-2959
