+++ kong-3.6.0-0.rockspec
@@ -200,6 +200,7 @@ build = {
     ["kong.workspaces"] = "kong/workspaces/init.lua",
 
     ["kong.db"] = "kong/db/init.lua",
+    ["kong.db.utils"] = "kong/db/utils.lua",
     ["kong.db.errors"] = "kong/db/errors.lua",
     ["kong.db.iteration"] = "kong/db/iteration.lua",
     ["kong.db.dao"] = "kong/db/dao/init.lua",
